SFX Manager
Bob, If you're reading this. This was harder than it looks.
Added the Maze countdown in audio.
Added SFX manager including Maps for SFX, Ambience Music, and walking noises.
Walking noises do work depending on ground walking noises are walking on.
Added functionality so that if we add a character model, the walking noises will happen when each foot of the character hits the ground. Which is very nice